Refactor: Change img tags to svg components
