+++ src/components/common/sideBar/DropDown.tsx
@@ -25,7 +25,7 @@ function DropDownItem({ children }: DropDownItemProps) {
     <button
       onClick={handleClickOpenModal}
       type="button"
-      className="text-body3-medium inline-flex h-[3.7rem] items-center justify-center p-4 text-gray80 hover:bg-[#EAEAEA]"
+      className="inline-flex h-[3.7rem] items-center justify-center p-4 text-body3-medium text-gray80 hover:bg-[#EAEAEA]"
     >
       {children}
     </button>
