+++ src/pages/MyIssuesPage.tsx
@@ -1,5 +1,7 @@
+import PageLayout from '@/components/common/PageLayout';
+
 const MyIssuesPage = () => {
-  return <div>여기는 이슈페이지입니다</div>;
+  return <PageLayout>{}</PageLayout>;
 };
 
 export default MyIssuesPage;
